# docs/openapi.yml
...
openapi: 3.1.0
components:
  schemas:
    ApplicationSpec:
      type: object
      deprecated: true
      required:
      - repoId
      - name
      properties:
        repoId:
          type: string
          pattern: \S
        name:
          type: string
          pattern: \S
        healthPath:
          type: string
    CreateEnvironmentRequest:
      type: object
      required:
      - name
      properties:
        name:
          type: string
          pattern: \S
        network:
          type: string
        designated:
          type: boolean
        platform:
          type: boolean
          deprecated: true
        applications:
          type: array
          deprecated: true
          items:
            $ref: "#/components/schemas/ApplicationSpec"
...
    EnvironmentResponse:
      type: object
      properties:
        environment:
          $ref: "#/components/schemas/PdEnvironmentDto"
    Instant:
      type: string
      format: date-time
      examples:
      - 2022-03-10T16:15:50Z
...
    ListEnvironmentsResponse:
      type: object
      properties:
        environments:
          type: array
          items:
            $ref: "#/components/schemas/PdEnvironmentDto"
...
    ListLinksResponse:
      type: object
      properties:
        services:
          type: array
          items:
            $ref: "#/components/schemas/PdLinkedServiceDto"
...
    PdApplicationDto:
      type: object
      properties:
        id:
          type: string
        repoId:
          type: string
        name:
          type: string
        environmentId:
          type: string
        environmentName:
          type: string
        availableOnEnv:
          type: boolean
        branch:
          type: string
        healthPath:
          type: string
        createdAt:
          $ref: "#/components/schemas/Instant"
...
    PdEnvironmentDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        network:
          type: string
        designated:
          type: boolean
        platform:
          type: boolean
        createdAt:
          $ref: "#/components/schemas/Instant"
        applications:
          type: array
          items:
            $ref: "#/components/schemas/PdApplicationDto"
...
    PdLinkedServiceDto:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        availableOnEnv:
          type: boolean
        healthPath:
          type: string
...
    UpdateEnvironmentRequest:
      type: object
      properties:
        name:
          type: string
        designated:
          type: boolean
        platform:
          type: boolean
          deprecated: true
...
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      openIdConnectUrl: http://dev-qits-idp:8080/idp/.well-known/openid-configuration
      description: Authentication
paths:
...
  /deployments/api/environments:
    get:
      summary: List environments
      operationId: listEnvironments
      responses:
        "200":
          description: The environments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListEnvironmentsResponse"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Pd Environment Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:agent
    post:
      summary: "Create an environment: a name and a bundle network"
      operationId: createEnvironment
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEnvironmentRequest"
        required: true
      responses:
        "201":
          description: Created; a release may now be deployed into it
        "400":
          description: A name or network failed validation
        "409":
          description: An environment of that name already exists
        "401":
          description: Gate on and no machine token presented
        "403":
          description: Gate on and the token is for another service
      tags:
      - Pd Environment Controller
      security:
      - SecurityScheme:
        - qits:system
  /deployments/api/environments/{environmentId}:
    patch:
      summary: "Rename an environment, or designate it the platform environment"
      operationId: updateEnvironment
      parameters:
      - name: environmentId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEnvironmentRequest"
        required: true
      responses:
        "200":
          description: The updated environment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvironmentResponse"
        "400":
          description: A name failed validation
        "404":
          description: No such environment
        "409":
          description: "Another environment already has that name, or the platform\
            \ designation was cleared rather than moved"
        "401":
          description: Gate on and no machine token presented
        "403":
          description: Gate on and the token is for another service
      tags:
      - Pd Environment Controller
      security:
      - SecurityScheme:
        - qits:system
    get:
      summary: One environment with the applications it tracks
      operationId: getEnvironment
      parameters:
      - name: environmentId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: The environment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvironmentResponse"
        "404":
          description: No such environment
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Pd Environment Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:agent
    delete:
      summary: "Tear an environment down (rows, containers, networks)"
      operationId: deleteEnvironment
      parameters:
      - name: environmentId
        in: path
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Torn down
        "404":
          description: No such environment
        "409":
          description: This is the platform environment
        "401":
          description: Gate on and no machine token presented
        "403":
          description: Gate on and the token is for another service
      tags:
      - Pd Environment Controller
      security:
      - SecurityScheme:
        - qits:system
  /deployments/api/environments/{environmentId}/links:
    get:
      summary: "Every service present in this environment: its links, plus every platform\
        \ service"
      operationId: listEnvironmentLinks
      parameters:
      - name: environmentId
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: The services present in this environment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListLinksResponse"
        "404":
          description: No such environment
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      tags:
      - Pd Environment Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
        - qits:agent
...
info:
  title: qits-platform-deployments API
  version: 2026.1008.164327